Handle the case where we have fewer unit test splits than workers
We were already attempting to handle this -- we would only look at the
results from workers that had tests to run.  But the workers who didn't
get any split at all would also crash; the way the unit test runner is
set up that fails the job even though we ignored the results.  Now, we
check for that and early-out.

In general we could probably do with sharing more code between e2e-test
and webapp-test; the former already had a fix for this problem.  They do
have some differences, of course.

Test plan: kevinb's tests should work

Auditors: csilvers, kevinb

+++ jobs/webapp-test.groovy
@@ -286,6 +286,13 @@ def doTestOnWorker(workerNum) {
       // get the same one, and start right away, once ready.
       waitUntil({ HAVE_STASHED_TESTS });
 
+      // It may be we turned out not to need this worker, because there were so
+      // few tests to run.  (NUM_WORKER_MACHINES will in that case get reset in
+      // _determineTests above.)
+      if (workerNum >= NUM_WORKER_MACHINES) {
+         return;
+      }
+
       // Out with the old, in with the new!
       sh("rm -f test-results.*.pickle");
       unstash("splits");
